Add files via upload
The data are the pre-processed fMRI time series out of anatomy-informed partition. Dim.txt is the dimension of the fMRI volume. 
Name_Number is the Region_ROI. For example, Fro_1.txt to Fro_9.txt are the data from 9 regions in left frontal lobe, and Fro_10.txt to Fro_18.txt are the data from 9 regions in right frontal lobe. The regions include frontal, limbic, temporal, parietal, occipital and other.